ultralytics_object/requirements.txt
paho-mqtt==1.5.1
onnxruntime_gpu==1.11.0
#onnx==1.10.0
urllib3==1.26.6
requests==2.25.1
numpy==1.22.1
opencv-python==4.8.0.76
minio==7.1.8
omegaconf